+++ client/src/containers/analysis-visualization/component.tsx
@@ -2,7 +2,8 @@ import { useAppSelector } from 'store/hooks';
 import { analysis } from 'store/features/analysis';
 
 import Table from './table';
-import { COLUMNS, DATA } from './table/mock';
+import DATA from './table/mock';
+import Title from './table/title';
 
 import DatasetControl from './dataset-control';
 import ModeControl from './mode-control';
@@ -12,6 +13,53 @@ import AnalysisFilters from './analysis-filters';
 
 const AnalysisVisualization: React.FC = () => {
   const { visualizationMode } = useAppSelector(analysis);
+  const TABLE_COLUMNS = [
+    {
+      title: 'YEAR',
+      dataIndex: 'commodity',
+      key: 'commodity',
+      width: 150,
+      fixed: 'left',
+    },
+    {
+      title: '2021-2025',
+      dataIndex: 'all',
+      key: 'all',
+      width: 100,
+      fixed: 'left',
+    },
+    {
+      title: () => <Title title="2021" />,
+      dataIndex: '2021',
+      key: '2021',
+      width: 100,
+      fixed: 'left',
+    },
+    {
+      title: () => <Title title="2022" />,
+      dataIndex: '2022',
+      key: '2022',
+      width: 100,
+    },
+    {
+      title: () => <Title title="2023" />,
+      dataIndex: '2023',
+      key: '2023',
+      width: 100,
+    },
+    {
+      title: () => <Title title="2024" />,
+      dataIndex: '2024',
+      key: '2024',
+      width: 100,
+    },
+    {
+      title: () => <Title title="2025" />,
+      dataIndex: '2025',
+      key: '2025',
+      width: 100,
+    },
+  ];
 
   return (
     <section className="relative flex flex-col flex-1 w-screen md:w-full sm:h-screen-minus-header md:h-full bg-gray-50 lg:order-last">
@@ -32,8 +80,8 @@ const AnalysisVisualization: React.FC = () => {
       {visualizationMode === 'map' && <AnalysisMap />}
 
       {visualizationMode === 'table' && (
-        <div className="flex flex-col p-6 pl-12 mt-16 left-12" style={{ width: '1500px' }}>
-          <Table columns={COLUMNS} dataSource={DATA} />
+        <div className="flex p-6 pl-12 mt-16 left-12" style={{ maxWidth: '1800px' }}>
+          <Table columns={TABLE_COLUMNS} dataSource={DATA} />
         </div>
       )}
