+++ client/src/containers/analysis-visualization/table/component.tsx
@@ -14,7 +14,7 @@ export interface TableProps {
 
 export const Table: React.FC<TableProps> = ({ columns, dataSource }: TableProps) => (
   <TableAntd
-    className="w-full"
+    className="max-w-max"
     expandable={{
       rowExpandable: (record) => !!record.children,
       expandIcon: ({ expanded, onExpand, record }) => {
