+++ client/src/containers/analysis-visualization/table/title/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
